New Version Policy (#752)
* Rename 1.20.6 module

* Rename 1.21.6

* Update version loading code

* Backport 1.21.6 to 1.21.5

* Backport 1.21.6 to 1.21.4

* Backport 1.21.6 to 1.21.3

* Backport 1.21.6 to 1.21.1

* Remove 1.21.3

Signs and inventories clear, not sure why
