+++ Movecraft/build.gradle.kts
@@ -8,8 +8,11 @@ plugins {
 java.toolchain.languageVersion = JavaLanguageVersion.of(21)
 
 dependencies {
-    runtimeOnly(project(":movecraft-v1_20", "reobf"))
-    runtimeOnly(project(":movecraft-v1_21", "reobf"))
+    runtimeOnly(project(":movecraft-v1_20_6", "reobf"))
+    runtimeOnly(project(":movecraft-v1_21_1", "reobf"))
+    runtimeOnly(project(":movecraft-v1_21_4", "reobf"))
+    runtimeOnly(project(":movecraft-v1_21_5", "reobf"))
+    runtimeOnly(project(":movecraft-v1_21_6", "reobf"))
     implementation(project(":movecraft-api"))
     compileOnly("org.yaml:snakeyaml:2.0")
 }
@@ -21,8 +24,11 @@ tasks.shadowJar {
 
     dependencies {
         include(project(":movecraft-api"))
-        include(project(":movecraft-v1_20"))
-        include(project(":movecraft-v1_21"))
+        include(project(":movecraft-v1_20_6"))
+        include(project(":movecraft-v1_21_1"))
+        include(project(":movecraft-v1_21_4"))
+        include(project(":movecraft-v1_21_5"))
+        include(project(":movecraft-v1_21_6"))
     }
 }
 
@@ -66,7 +72,7 @@ hangarPublish {
         platforms {
             register(io.papermc.hangarpublishplugin.model.Platforms.PAPER) {
                 jar.set(tasks.shadowJar.flatMap { it.archiveFile })
-                platformVersions.set(listOf("1.20.6", "1.21.5"))
+                platformVersions.set(listOf("1.20.6", "1.21.1", "1.21.4", "1.21.5", "1.21.6"))
             }
         }
     }
