+++ build.gradle.kts
@@ -1,7 +1,3 @@
-/*
- * This file was generated by the Gradle 'init' task.
- */
-
 plugins {
     id("buildlogic.java-conventions")
     id("io.papermc.paperweight.userdev") version "2.0.0-beta.17" apply false
