+++ settings.gradle.kts
@@ -1,15 +1,17 @@
-/*
- * This file was generated by the Gradle 'init' task.
- */
-
 rootProject.name = "movecraft-parent"
-include(":movecraft-v1_20")
-include(":movecraft-v1_21")
+include(":movecraft-v1_20_6")
+include(":movecraft-v1_21_1")
+include(":movecraft-v1_21_4")
+include(":movecraft-v1_21_5")
+include(":movecraft-v1_21_6")
 include(":movecraft-api")
 include(":movecraft-datapack")
 include(":movecraft")
-project(":movecraft-v1_20").projectDir = file("v1_20")
-project(":movecraft-v1_21").projectDir = file("v1_21")
+project(":movecraft-v1_20_6").projectDir = file("v1_20_6")
+project(":movecraft-v1_21_1").projectDir = file("v1_21_1")
+project(":movecraft-v1_21_4").projectDir = file("v1_21_4")
+project(":movecraft-v1_21_5").projectDir = file("v1_21_5")
+project(":movecraft-v1_21_6").projectDir = file("v1_21_6")
 project(":movecraft-api").projectDir = file("api")
 project(":movecraft-datapack").projectDir = file("datapack")
 project(":movecraft").projectDir = file("Movecraft")
