--- v1_20/build.gradle.kts
@@ -1,18 +1,0 @@
-/*
- * This file was generated by the Gradle 'init' task.
- */
-
-plugins {
-    id("buildlogic.java-conventions")
-    id("io.papermc.paperweight.userdev")
-}
-
-java.toolchain.languageVersion = JavaLanguageVersion.of(21)
-
-dependencies {
-    api(project(":movecraft-api"))
-    paperweight.paperDevBundle("1.20.6-R0.1-SNAPSHOT")
-}
-
-description = "Movecraft-v1_20"
-paperweight.reobfArtifactConfiguration = io.papermc.paperweight.userdev.ReobfArtifactConfiguration.MOJANG_PRODUCTION
